[RPC] Break out testing utils
Closes: #8435
