+++ json-rpc/src/tests/mod.rs
@@ -2,11 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(test)]
-mod genesis;
-#[cfg(test)]
 mod unit_tests;
-mod utils;
 
-pub use utils::test_bootstrap;
 #[cfg(any(test, feature = "fuzzing"))]
-pub use utils::MockDiemDB;
+pub(crate) mod genesis;
+
+#[cfg(any(test, feature = "fuzzing"))]
+pub(crate) mod utils;
